net/mlx5e: Extract encap entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up encap entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new mod header entry into standalone function as a preparation for
described refactoring.

issue: 1521224
Change-Id: Ie172266b4d832c6364a00b82c97db60a24408d75
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>
Reviewed-by: Jianbo Liu <jianbol@mellanox.com>