net/mlx5e: eswitch_offloads: Add chains and priorities
A chain is a group of priorities, so use the fdb parallel
sub namespaces to implement chains, and a flow table for each
priority in them.

Because these namespaces are parallel and in series to slow path fdb,
the chains aren't connecteded to one another (but to the slow path),
and one must use a explicit goto action to reach a new chain.

Flow tables for the priorities will be created on demand
and destroyed once not used.

This isn't enabled yet by this patch, for now all flows will
use the default chain 0, and prio 1.

Issue: 1323489
Change-Id: If81a3e7d576e8007c3a32d65eb2f8477420119e9
Signed-off-by: Paul Blakey <paulb@mellanox.com>