net/mlx5: Fix duplicated offload rules issue
The offending commit fixes the tc replace issue. Now it exposes
another issue. If user adds two same filters using skip_sw, the
second command will succeed.

In this patch, if the matches are the same, we will compare the
filter handles. If the handles are the same, we consider it as
a replace command and continue. If the handles are not the same,
we consider it as a duplicate add and fail the command.

Issue: 1547947
Change-Id: Iac3269a710e0cfdec66e5acb3882b60e52c81a6f
Fixes: 42f7ad676007 ("net/mlx5e: For TC offloads, always add new flow instead of appending the actions")
Signed-off-by: Chris Mi <chrism@mellanox.com>

Conflicts:
	drivers/net/ethernet/mellanox/mlx5/core/en_tc.c